Move LIBS after the OBJS
According to the ld manpage
http://sourceware.org/binutils/docs/ld/Options.html#Options

The linker will search an archive only once, at the location where it is specified on the command line. If the archive defines a symbol which was undefined in some object which appeared before the archive on the command line, the linker will include the appropriate file(s) from the archive. However, an undefined symbol in an object appearing later on the command line will not cause the linker to search the archive again.